+++ packages/core/src/secrets/index.ts
@@ -2,5 +2,5 @@ export * from './SecretManager.js';
 export * from './loaders/index.js';
 export * from './providers/index.js';
 export * from './types.js';
-export * from './SecretsOrchestrator.js';
+export * from './orchestrator/index.js';
 export * from './SecretsInjectionContext.js';
