+++ packages/core/src/secrets/orchestrator/index.ts
@@ -1,0 +1,1 @@
+export * from './SecretsOrchestrator.js';
