+++ packages/core/src/types.ts
@@ -1,5 +1,6 @@
 import type { BaseStack } from './BaseStack.js';
 import { ResourceComposer } from './ResourceComposer.js';
+import type { MergeLevel, MergeStrategy } from './secrets/index.js';
 
 export type FunctionLike<Params extends unknown[] = [], Return = unknown> = (...args: Params) => Return;
 export type AnyFunction = FunctionLike<unknown[], unknown>;
@@ -30,8 +31,14 @@ export type InferConfigureComposerFunc<T> = T extends (...args: any[]) => Resour
  */
 export type AnyString = string & {};
 
+
 export interface KubricateConfig {
   stacks?: Record<string, BaseStack>;
+  secrets?: {
+    kubernetes?: {
+      merge?: Partial<Record<MergeLevel, MergeStrategy>>;
+    }
+  }
 }
 
 export type LogLevel = 'silent' | 'error' | 'warn' | 'info' | 'debug';
